% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destructure.R
\name{deconstruct}
\alias{deconstruct}
\title{Deconstruct a dataframe or list into vectors available in the global environment}
\usage{
deconstruct(x)
}
\arguments{
\item{x}{a dataframe or list to deconstruct}
}
\description{
Deconstruct a dataframe or list into vectors available in the global environment
}
\examples{
deconstruct(mtcars)
}
