% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpreter.R
\name{interpreter}
\alias{interpreter}
\title{Function generator for constructing interpretations}
\usage{
interpreter(stnc)
}
\arguments{
\item{stnc}{The sentence to be used for constructing interpretations. Values are dynamically changed through \code{glue::glue()}.}
}
\value{
A character string of the constructed interpretation.
}
\description{
Function generator for constructing interpretations
}
\examples{
new_interpretation <- interpreter("My p-value is {P}")
new_interpretation(P = 0.03)
}
