% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promis.R
\name{label_promis}
\alias{label_promis}
\title{Apply clinical labels to PROMIS scores}
\usage{
label_promis(domain, score)
}
\arguments{
\item{domain}{the PROMIS domain, a string.}

\item{score}{the appropriate PROMIS domain score, a scalar.}
}
\value{

}
\description{
Apply clinical labels to PROMIS scores
}
\examples{
label_promis("physfx", 42)
}
