% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promis.R
\name{pretty_promis_labels}
\alias{pretty_promis_labels}
\title{Create pretty PROMIS labels}
\usage{
pretty_promis_labels(domain)
}
\arguments{
\item{domain}{the PROMIS domain, a string.}
}
\value{
the prettified domain lable, a string.
}
\description{
Create pretty PROMIS labels
}
\examples{
pretty_promis_labels("physfx")
}
