% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promis.R
\name{promis_domains}
\alias{promis_domains}
\title{Function factory for PROMIS score clinical labels}
\usage{
promis_domains(domain)
}
\arguments{
\item{domain}{the PROMIS domain, a string.}
}
\value{
a function that takes as an input a PROMIS score and outputs the clinical interpretation for the specific domain.
}
\description{
Function factory for PROMIS score clinical labels
}
\examples{
promis_domains("physfx")(42)
}
