% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{read_excel}
\alias{read_excel}
\title{Low-level interface for reading excel workbooks}
\usage{
read_excel(
  path,
  sheet = NULL,
  rows = NULL,
  columns = NULL,
  na = "",
  trim_ws = TRUE,
  progress = readxl_progress(),
  .name_repair = "unique"
)
}
\arguments{
\item{.name_repair}{}
}
\value{

}
\description{
Low-level interface for reading excel workbooks
}
