% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtsummary_options.R
\name{wcmtheme.gtsummary_options}
\alias{wcmtheme.gtsummary_options}
\title{Set gtsummary options for wcmtheme}
\usage{
wcmtheme.gtsummary_options()
}
\description{
Set gtsummary options for wcmtheme
}
