% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatix.R
\name{wrap_parenthesis}
\alias{wrap_parenthesis}
\title{Wrap text in paranthesis}
\usage{
wrap_parenthesis(x)
}
\arguments{
\item{x}{a character or numeric value.}
}
\value{
a character representation of the original input surrounded in parenthesis.
}
\description{
Wrap text in paranthesis
}
\examples{
wrap_parenthesis(55)
}
